Option to set sale order done consider shipping
The option to automatically set the sales order as 'done' currently
filters only the sales orders in state 'sale' and invoiced.

I propose to change the default filter to 'all is invoiced AND all is
delivered' which emulates the behaviour of the previous versions of
Odoo.

This is a backward compatible change: as the filter has `noupdate="1"`,
existing installations will keep their domains and new installations
will get the new one.