[13.0][FIX] sale_exception ignore after confirm
After the sale order has been confirmed, some fields can still be
changed, those changes can raise some more excetptions again.
Unfortunatelly those changes are raised as ValidationError and they are
blocking.

If during the confirmation of the sale.order the ignore excetpiton has
been used. Those same ignored exception will block some changes made
when the sale order is in `sale` state.

And this problem is due to that onchange that this PR removes.
Why should the ignore_exception be set to false when lines are changed
but not when some fields on the sale.order record are changed ?