[IMP] sale_order_product_recommendation: choose order on mobile
Before, recommendations on mobile screens couldn't be sorted because they are using the kanban widget, which doesn't allow it.

Recommendation lines order is particularly important to improve salesperson efficiency, so a special feature is added here so that it can be done while using the module in mobile.

The previous default behavior is respected.

@moduon MT-4472