[IMP] sale_order_product_recommendation: sort separately by product name or code
Before this patch, only one column was displayed for product. Thus, sorting by its name was giving unexpected results if some had code and others didn't.

Now that column is hidden by default, and product name and code are split in 2 columns that can be sorted separately in list view.

As the amount of list columns grows, I added `optional="show"` to more of them.

For mobile kanban, product code order is added to the settings before generating recommendations.

@moduon MT-4472