[FIX] sale_order_product_recommendation: make UoM work
Before this patch, UoM wasn't being properly inherited into generated lines when those lines didn't exist previously in the SO.

Also, changing it manually didn't work as expected in that same case.

This is because the UoM was related to the SOL UoM. But if there was no SOL, it did nothing.

Now it's properly computed. UoM comes from the product if there's no SOL or it can't provide it. The user can change it to other UoM of the same category. The change is propagated when applying the wizard, both when creating lines and updating them. Tests added.

@moduon MT-4472