[FIX] sale_order_type: Incompatibility with l10n_latam_check module
This is a work-in-progress PR and is not yet finished.

When adding a dependency to the _compute_journal_id method, it breaks the functionality of the l10n_latam_check module when calculating the l10n_latam_check_issuer_vat field during the creation of a third-party check payment.

Steps to reproduce:

Create a payment.
Choose a third-party check journal.
Modify the l10n_latam_check_issuer_vat issuer.
Save the payment.
The l10n_latam_check_issuer_vat field is recomputed.